Extract.
